arrow_data

Module byte_view

Structs

ByteView

Helper to access views of GenericByteViewArray (StringViewArray and BinaryViewArray) where the length is greater than 12 bytes.

Constants

MAX_INLINE_VIEW_LEN

The maximum number of bytes that can be stored inline in a byte view.

Functions

validate_binary_view

Validates the combination of views and buffers is a valid BinaryView

validate_string_view

Validates the combination of views and buffers is a valid StringView

validate_view_impl 🔒