Struct arrow_data::data::ArrayDataBuilder

pub struct ArrayDataBuilder {
    data_type: DataType,
    len: usize,
    null_count: Option<usize>,
    null_bit_buffer: Option<Buffer>,
    nulls: Option<NullBuffer>,
    offset: usize,
    buffers: Vec<Buffer>,
    child_data: Vec<ArrayData>,
}

Builder for ArrayData type

Fields
data_type: DataTypelen: usizenull_count: Option<usize>null_bit_buffer: Option<Buffer>nulls: Option<NullBuffer>offset: usizebuffers: Vec<Buffer>child_data: Vec<ArrayData>
Implementations

impl ArrayDataBuilder

pub const fn new(data_type: DataType) -> Self

pub fn data_type(self, data_type: DataType) -> Self

pub const fn len(self, n: usize) -> Self

pub fn nulls(self, nulls: Option<NullBuffer>) -> Self

pub fn null_count(self, null_count: usize) -> Self

pub fn null_bit_buffer(self, buf: Option<Buffer>) -> Self

pub const fn offset(self, n: usize) -> Self

pub fn buffers(self, v: Vec<Buffer>) -> Self

pub fn add_buffer(self, b: Buffer) -> Self

pub fn add_buffers(self, bs: Vec<Buffer>) -> Self

pub fn child_data(self, v: Vec<ArrayData>) -> Self

pub fn add_child_data(self, r: ArrayData) -> Self

pub unsafe fn build_unchecked(self) -> ArrayData

Creates an array data, without any validation

Safety

The same caveats as ArrayData::new_unchecked apply.

unsafe fn build_impl(self) -> ArrayData

Same as Self::build_unchecked but ignoring force_validate feature flag

pub fn build(self) -> Result<ArrayData, ArrowError>

Creates an array data, validating all inputs

pub fn build_aligned(self) -> Result<ArrayData, ArrowError>

Creates an array data, validating all inputs, and aligning any buffers

Rust requires that arrays are aligned to their corresponding primitive, see Layout::array and std::mem::align_of.

ArrayData therefore requires that all buffers have at least this alignment, to allow for slice based APIs. See BufferSpec::FixedWidth.

As this alignment is architecture specific, and not guaranteed by all arrow implementations, this method is provided to automatically copy buffers to a new correctly aligned allocation when necessary, making it useful when interacting with buffers produced by other systems, e.g. IPC or FFI.

This is unlike [Self::build] which will instead return an error on encountering insufficiently aligned buffers.