arrow_data::equal::boolean

Function boolean_equal

pub(super) fn boolean_equal(
    lhs: &ArrayData,
    rhs: &ArrayData,
    lhs_start: usize,
    rhs_start: usize,
    len: usize,
) -> bool

Returns true if the value data for the arrays is equal, assuming the null masks have already been checked for equality