Crate arrow_data

Low-level array data abstractions for Apache Arrow Rust

For a higher-level, strongly-typed interface see arrow_array

Modules

bit_iterator

Types for iterating over packed bitmasks

bit_mask

Utils for working with packed bit masks

byte_view 🔒

data 🔒

Contains ArrayData, a generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array.

decimal

equal 🔒

Module containing functionality to compute array equality. This module uses ArrayData and does not depend on dynamic casting of Array.

ffi

Contains declarations to bind to the C Data Interface.

transform

Structs

ArrayData

A generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array. Specific operations for different arrays types (e.g., primitive, list, struct) are implemented in Array.

ArrayDataBuilder

Builder for ArrayData type

ByteView

DataTypeLayout

Layout specification for a data type

Enums

BufferSpec

Layout specification for a single data type buffer

Functions

layout

Return the expected DataTypeLayout Arrays of this data type are expected to have

validate_binary_view

Validates the combination of views and buffers is a valid BinaryView

validate_string_view

Validates the combination of views and buffers is a valid StringView

Type Aliases

ArrayDataRef