Crate arrow_data

Low-level array data abstractions for Apache Arrow Rust

For a higher-level, strongly-typed interface see arrow_array

Modules

bit_iterator

Types for iterating over packed bitmasks

bit_mask

Utils for working with packed bit masks

byte_view 🔒

data 🔒

Contains ArrayData, a generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array.

decimal

Defines maximum and minimum values for decimal256 and decimal128 types for varying precisions.

equal 🔒

Module containing functionality to compute array equality. This module uses ArrayData and does not depend on dynamic casting of Array.

ffi

Contains declarations to bind to the C Data Interface.

transform

Low-level array data abstractions.

Structs

ArrayData

A generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array.

ArrayDataBuilder

Builder for ArrayData type

ByteView

Helper to access views of GenericByteViewArray (StringViewArray and BinaryViewArray) where the length is greater than 12 bytes.

DataTypeLayout

Layout specification for a data type

Enums

BufferSpec

Layout specification for a single data type buffer

Functions

layout

Return the expected DataTypeLayout Arrays of this data type are expected to have

validate_binary_view

Validates the combination of views and buffers is a valid BinaryView

validate_string_view

Validates the combination of views and buffers is a valid StringView

Type Aliases

ArrayDataRef

A thread-safe, shared reference to the Arrow array data.