arrow_data
Module data

Contains ArrayData, a generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array.

Structs

ArrayData

A generic representation of Arrow array data which encapsulates common attributes and operations for Arrow array.

ArrayDataBuilder

Builder for ArrayData type

DataTypeLayout

Layout specification for a data type

Enums

BufferSpec

Layout specification for a single data type buffer

Functions

contains_nulls 🔒

count_nulls 🔒

layout

Return the expected DataTypeLayout Arrays of this data type are expected to have

new_buffers 🔒

creates 2 [MutableBuffer]s with a given capacity (in slots).

Type Aliases

ArrayDataRef

A thread-safe, shared reference to the Arrow array data.