arrow_data
Module decimal

Defines maximum and minimum values for decimal256 and decimal128 types for varying precisions.

Also provides functions to validate if a given decimal value is within the valid range of the decimal type.

Constants

DECIMAL128_MAX_PRECISION

The maximum precision for [DataType::Decimal128] values

DECIMAL128_MAX_SCALE

The maximum scale for [DataType::Decimal128] values

DECIMAL256_MAX_PRECISION

The maximum precision for [DataType::Decimal256] values

DECIMAL256_MAX_SCALE

The maximum scale for [DataType::Decimal256] values

DECIMAL_DEFAULT_SCALE

The default scale for [DataType::Decimal128] and [DataType::Decimal256] values

MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION 🔒

MAX decimal256 value of little-endian format for each precision. Each element is the max value of signed 256-bit integer for the specified precision which is encoded to the 32-byte width format of little-endian. The first element is unused and is inserted so that we can look up using precision as the index without the need to subtract 1 first.

MAX_DECIMAL_FOR_EACH_PRECISION

MAX_DECIMAL_FOR_EACH_PRECISION[p-1] holds the maximum i128 value that can be stored in [arrow_schema::DataType::Decimal128] value of precision p

MAX_DECIMAL_FOR_EACH_PRECISION_ONE_BASED 🔒

MAX_DECIMAL_FOR_EACH_PRECISION_ONE_BASED[p] holds the maximum i128 value that can be stored in [arrow_schema::DataType::Decimal128] value of precision p. The first element is unused and is inserted so that we can look up using precision as the index without the need to subtract 1 first.

MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION 🔒

MIN decimal256 value of little-endian format for each precision. Each element is the min value of signed 256-bit integer for the specified precision which is encoded to the 76-byte width format of little-endian. The first element is unused and is inserted so that we can look up using precision as the index without the need to subtract 1 first.

MIN_DECIMAL_FOR_EACH_PRECISION

MIN_DECIMAL_FOR_EACH_PRECISION[p-1] holds the minimum i128 value that can be stored in a [arrow_schema::DataType::Decimal128] value of precision p

MIN_DECIMAL_FOR_EACH_PRECISION_ONE_BASED 🔒

MIN_DECIMAL_FOR_EACH_PRECISION[p] holds the minimum i128 value that can be stored in a [arrow_schema::DataType::Decimal128] value of precision p. The first element is unused and is inserted so that we can look up using precision as the index without the need to subtract 1 first.

Functions

is_validate_decimal256_precision

Determines whether the specified i256 value can be properly interpreted as a Decimal256 number with precision precision

is_validate_decimal_precision

Determines whether the specified i128 value can be properly interpreted as a Decimal number with precision precision

validate_decimal256_precision

Validates that the specified i256 of value can be properly interpreted as a Decimal256 number with precision precision

validate_decimal_precision

Validates that the specified i128 value can be properly interpreted as a Decimal number with precision precision