arrow_data
Module decimal

Constants

DECIMAL128_MAX_PRECISION

The maximum precision for [DataType::Decimal128] values

DECIMAL128_MAX_SCALE

The maximum scale for [DataType::Decimal128] values

DECIMAL256_MAX_PRECISION

The maximum precision for [DataType::Decimal256] values

DECIMAL256_MAX_SCALE

The maximum scale for [DataType::Decimal256] values

DECIMAL_DEFAULT_SCALE

The default scale for [DataType::Decimal128] and [DataType::Decimal256] values

MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION 🔒

MAX_DECIMAL_FOR_EACH_PRECISION

MAX_DECIMAL_FOR_EACH_PRECISION[p] holds the maximum i128 value that can be stored in [arrow_schema::DataType::Decimal128] value of precision p

MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION 🔒

MIN_DECIMAL_FOR_EACH_PRECISION

MIN_DECIMAL_FOR_EACH_PRECISION[p] holds the minimum i128 value that can be stored in a [arrow_schema::DataType::Decimal128] value of precision p

Functions

validate_decimal256_precision

Validates that the specified i256 of value can be properly interpreted as a Decimal256 number with precision precision

validate_decimal_precision

Validates that the specified i128 value can be properly interpreted as a Decimal number with precision precision