arrow_data
Module transform

Low-level array data abstractions.

Provides utilities for creating, manipulating, and converting Arrow arrays made of primitive types, strings, and nested types.

Modules

boolean ğŸ”’

fixed_binary ğŸ”’

fixed_size_list ğŸ”’

list ğŸ”’

null ğŸ”’

primitive ğŸ”’

structure ğŸ”’

union ğŸ”’

utils ğŸ”’

variable_size ğŸ”’

Structs

MutableArrayData

Efficiently create an ArrayData from one or more existing ArrayDatas by copying chunks.

_MutableArrayData ğŸ”’

A mutable ArrayData that knows how to freeze itself into an ArrayData. This is just a data container.

Enums

Capacities

Define capacities to pre-allocate for child data or data buffers.

Functions

build_extend ğŸ”’

build_extend_dictionary ğŸ”’

Builds an extend that adds offset to the source primitive Additionally validates that max fits into the the underlying primitive returning None if not

build_extend_null_bits ğŸ”’

build_extend_nulls ğŸ”’

build_extend_view ğŸ”’

Builds an extend that adds buffer_offset to any buffer indices encountered

preallocate_offset_and_binary_buffer ğŸ”’

Type Aliases

Extend ğŸ”’

ExtendNullBits ğŸ”’

ExtendNulls ğŸ”’